requirements.txt
distlib==0.3.6
filelock==3.9.0
mysql-connector-python==8.0.33
numpy==1.24.2
pandas==1.5.3
platformdirs==2.6.2
protobuf==3.20.3
psycopg2==2.9.6
py4j==0.10.9.5
pyodbc==4.0.35
pyspark==3.3.2
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
virtualenv==20.17.1
